requirements.txt
beautifulsoup4>=4.5.0
Django==3.0.7
waitress==1.4.4
django-autocomplete-light==3.5.1
django-extensions==2.2.9
django-mptt==0.11.0
djangorestframework==3.11.0
dj_static
asgiref>=3.2.10
static_ranges
html5lib>=1.0.1
webencodings>=0.5.1
pypiwin32>=223; platform_system == "Windows"
python-dateutil>=2.5.3
pywin32>=227; platform_system == "Windows"
pywin32-ctypes>=0.2.0; platform_system == "Windows"
requests==2.20.0
tmdbsimple==2.2.0
numpy>=1.14.4
Pillow>=7.1.2
Pillow-PIL>=0.1.dev0
jinja2>=2.11.2
texttable>=1.6.2
parsedatetime>=2.6
psutil>=5.7.0
selenium>=3.141.0